Remove plugin system — flatten meetings, briefings, activity log into core
The plugin abstraction added indirection without benefit for a codebase used
as a teaching demo. Inlining all features into direct code paths makes the
app easier for both humans and AI to modify correctly on the first try.
