--- app/plugins/briefings/schema.ts
@@ -1,16 +1,0 @@
-import { pgTable, text, serial, integer, timestamp } from "drizzle-orm/pg-core";
-import { createInsertSchema } from "drizzle-zod";
-import { z } from "zod";
-import { contacts } from "@shared/schema";
-
-export const briefings = pgTable("briefings", {
-  id: serial("id").primaryKey(),
-  contactId: integer("contact_id").notNull().references(() => contacts.id, { onDelete: "cascade" }),
-  content: text("content").notNull(),
-  createdAt: timestamp("created_at").notNull().defaultNow(),
-  updatedAt: timestamp("updated_at").notNull().defaultNow(),
-});
-
-export const insertBriefingSchema = createInsertSchema(briefings).omit({ id: true, createdAt: true, updatedAt: true });
-export type InsertBriefing = z.infer<typeof insertBriefingSchema>;
-export type Briefing = typeof briefings.$inferSelect;
